fix missing rename of vtiger custom

+++ docker/script/reset.sh
@@ -2,14 +2,14 @@
 
 echo "===> You are about to erase everything (type in capital letter: YES)?: "
 
-#read AGREE
+read AGREE
 
-#if [ "${AGREE}" = "YES" ]; then
-  docker-compose down -v --remove-orphans
-  docker-compose run --rm --no-deps myddleware rm -fr vendor node_modules
-  docker-compose run --rm --no-deps myddleware chmod -R 777 docker/tmp public/build public/js var/cache/prod docker/var/mysql
+if [ "${AGREE}" = "YES" ]; then
+  docker-compose -f docker/env/script.yml down -v --remove-orphans
+  docker-compose -f docker/env/script.yml run --rm --no-deps myddleware rm -fr vendor node_modules
+  docker-compose -f docker/env/script.yml run --rm --no-deps myddleware chmod -R 777 docker/tmp public/build public/js var/cache/prod docker/var/mysql
   mkdir -p docker/var/mysql && touch docker/var/mysql/.gitkeep
   git clean -dfx
-#else
-#  echo "Wrong response!"
-#fi
+else
+  echo "Wrong response!"
+fi
